init script update

+++ rootfs/opt/init.sh
@@ -7,10 +7,7 @@ echo "Grabbing PUID=${PUID} and PGID=${PGID}..."
 echo "Grabbing UMASK=${UMASK}..."
 echo "Grabbing DELUGE_LOGLEVEL=${DELUGE_LOGLEVEL}..."
 echo "Grabbing TZ=${TZ}..."
-
 echo "Grabbing PYTHON_VER=${PYTHON_VER}..."
-readonly PLUGIN_NAME=$(ls /defaults/plugins | cut -c 1-20)
-echo "Grabbing PLUGIN_NAME=${PLUGIN_NAME}..."
 
 # Change IDs of xyz user and group
 echo "Setting PUID=${PUID} and PGID=${PGID} to xyz user and group"
@@ -20,13 +17,17 @@ usermod -o -u "$PUID" xyz
 # Handling default configs
 echo "Handling default config for Deluge..."
 if [[ ! -f $DELUGE_CONFIG/core.conf ]]; then
-	cp /defaults/core.conf $DELUGE_CONFIG/core.conf
+	cp /defaults/*.conf $DELUGE_CONFIG/
 fi
 echo "Installing the FileBot plugin if it's missing..."
 if [[ ! -f $DELUGE_CONFIG/plugins/FileBotTool-*$PYTHON_VER.egg ]]; then
 	echo "FileBot plugin is missing... Installing..."
-	rm $DELUGE_CONFIG/plugins/FileBotTool-*.egg
-	cp /defaults/plugins/FileBotTool-*.egg $DELUGE_CONFIG/plugins/$PLUGIN_NAME$PYTHON_VER.egg
+	if [[ ! -f $DELUGE_CONFIG/plugins/FileBotTool-*.egg ]]; then
+		mkdir -p $DELUGE_CONFIG/plugins
+	else
+		rm $DELUGE_CONFIG/plugins/FileBotTool-*.egg
+	fi
+	cp /defaults/plugins/FileBotTool-*.egg $DELUGE_CONFIG/plugins/
 fi
 
 # Setting folder ownership
